backend/requirements.txt
# 웹 프레임워크
fastapi==0.115.6
uvicorn[standard]==0.32.1

# DB
asyncpg==0.30.0
psycopg2-binary==2.9.10
sqlmodel==0.0.22
alembic==1.14.0
python-dotenv==1.0.1

# 설정
pydantic-settings==2.7.0

# 인증
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# 엑셀 파싱
pandas==2.2.3
openpyxl==3.1.5

# 유틸리티
python-multipart==0.0.20
